# docs/superadmin-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/superadmin/support": {
      "get": {
        "operationId": "SupportController_listTickets",
        "parameters": [
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "description": "Filter by tenant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "IN_PROGRESS",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search subject or message for a substring",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupportTicketDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List support tickets",
        "tags": [
          "Support"
        ]
      },
      "post": {
        "operationId": "SupportController_createTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a support ticket for a tenant",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/superadmin/support/{id}": {
      "get": {
        "operationId": "SupportController_getTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get support ticket details",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/superadmin/support/{id}/status": {
      "patch": {
        "operationId": "SupportController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update support ticket status",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/superadmin/support/{id}/assign": {
      "patch": {
        "operationId": "SupportController_assignAgent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Assign agent to support ticket",
        "tags": [
          "Support"
        ]
      }
    },
    "/api/v1/superadmin/templates": {
      "get": {
        "operationId": "TemplatesController_listTemplates",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "METRIC",
                "OUTCOME",
                "ACTIVITY"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List global templates",
        "tags": [
          "Templates"
        ]
      },
      "post": {
        "operationId": "TemplatesController_createTemplate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new template",
        "tags": [
          "Templates"
        ]
      }
    },
    "/api/v1/superadmin/templates/{id}": {
      "get": {
        "operationId": "TemplatesController_getTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get template detail",
        "tags": [
          "Templates"
        ]
      },
      "patch": {
        "operationId": "TemplatesController_updateTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update template metadata",
        "tags": [
          "Templates"
        ]
      },
      "delete": {
        "operationId": "TemplatesController_deleteTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Deactivate template",
        "tags": [
          "Templates"
        ]
      }
    },
    "/api/v1/superadmin/templates/{id}/duplicate": {
      "post": {
        "operationId": "TemplatesController_duplicateTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Duplicate an existing template",
        "tags": [
          "Templates"
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Accountability SuperAdmin API",
    "description": "REST API for the BG Accountability Platform - SuperAdmin Services.\n\n## Overview\nThis API provides administrative capabilities for managing the multi-tenant platform.\nAccess is restricted to users with the `SUPER_ADMIN` role.\n\n## Authentication\nAll endpoints (except login) require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\nSuperAdmin accounts require MFA for enhanced security.\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n- MFA can be managed via `/api/v1/auth/mfa/*` endpoints\n\n## Session Management\n- View active sessions: `GET /api/v1/auth/sessions`\n- Revoke sessions: `DELETE /api/v1/auth/sessions/:id`\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token (SUPER_ADMIN role required)"
      }
    },
    "schemas": {
...
      "SupportTicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ticket_123"
          },
          "tenantId": {
            "type": "string",
            "example": "tenant_123"
          },
          "subject": {
            "type": "string",
            "example": "API outage"
          },
          "message": {
            "type": "string",
            "example": "Users cannot submit forms"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ],
            "example": "OPEN"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "MEDIUM"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-06-01T12:34:56.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-06-01T13:00:00.000Z"
          }
        },
        "required": [
          "id",
          "tenantId",
          "subject",
          "message",
          "status",
          "priority",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateSupportTicketDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "Tenant identifier",
            "example": "tenant_123"
          },
          "subject": {
            "type": "string",
            "description": "Ticket subject",
            "example": "Unable to access dashboard"
          },
          "message": {
            "type": "string",
            "description": "Detailed message",
            "example": "Dashboard throws 500 when loading metrics"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "default": "MEDIUM"
          }
        },
        "required": [
          "tenantId",
          "subject",
          "message"
        ]
      },
      "UpdateSupportStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "IN_PROGRESS",
              "RESOLVED"
            ],
            "example": "IN_PROGRESS"
          }
        },
        "required": [
          "status"
        ]
      },
      "AssignAgentDto": {
        "type": "object",
        "properties": {
          "assignedTo": {
            "type": "string",
            "description": "Agent name or ID to assign",
            "example": "support-agent-1"
          }
        },
        "required": [
          "assignedTo"
        ]
      },
      "TemplateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "template_123"
          },
          "type": {
            "type": "string",
            "enum": [
              "METRIC",
              "OUTCOME",
              "ACTIVITY"
            ],
            "example": "METRIC"
          },
          "name": {
            "type": "string",
            "example": "Weekly Revenue Metric"
          },
          "description": {
            "type": "object",
            "example": "Tracks weekly revenue numbers",
            "nullable": true
          },
          "payload": {
            "type": "object",
            "example": {
              "fields": [
                "amount",
                "notes"
              ]
            }
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-06-01T12:34:56.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-06-02T09:00:00.000Z"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "description",
          "payload",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTemplateDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "METRIC",
              "OUTCOME",
              "ACTIVITY"
            ],
            "example": "METRIC"
          },
          "name": {
            "type": "string",
            "description": "Template display name",
            "example": "Weekly Revenue Metric"
          },
          "description": {
            "type": "string",
            "description": "Optional description",
            "example": "Tracks revenue submitted every week"
          },
          "payload": {
            "type": "object",
            "description": "JSON payload describing template structure",
            "example": {
              "fields": [
                "amount",
                "notes"
              ]
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether template is active",
            "default": true
          }
        },
        "required": [
          "type",
          "name",
          "payload"
        ]
      },
      "UpdateTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Template display name"
          },
          "description": {
            "type": "string",
            "description": "Template description"
          },
          "payload": {
            "type": "object",
            "description": "Template payload",
            "example": {
              "fields": [
                "amount"
              ]
            }
          },
          "isActive": {
            "type": "boolean",
            "description": "Toggle template availability"
          }
        }
      },
...
    }
  }
}